routes and readme updated
